feat: moving the doProgressReport
